news-agent/pyproject.toml
[project]
name = "news-agent"
version = "0.1.0"
description = "A news agent that uses OpenAI's GPT-3 to generate news in a structured way, but all news are made up."
authors = [
    {name = "louislepage",email = "hello@louislepage.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "atomic-agents (>=1.0.19,<2.0.0)",
    "openai (>=1.61.1,<2.0.0)",
    "instructor (>=1.7.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]
[tool.poetry.scripts]
news-agent = "news_agent.__main__:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
